More neg strand test for star module
One of the tests for the neg strand differed from the pos strand in an
interesting way. When making the annotation dataframe while making the
SJ out panel, the annotation dataframe comes back in an unpredictable
order. This isn't an issue as I don't expect any particular order, but
if I make changes in the future that change the order, it will break my
tests. I could sort the annotation dataframe (and perhaps others to make
predictable order as well?) but I'd either have to do a typical numeric
or natural sort (ugly) or rely on pybedtools or something for a smarter
sort. Also, sorting could add performance costs although I don't think
that the dataframes that would be sorted are big enough to cause any
real damage.